Move to server-common code
