+++ core/src/main/java/kafka/server/share/SharePartition.java
@@ -32,7 +32,6 @@ import org.apache.kafka.common.message.ShareFetchResponseData.AcquiredRecords;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.record.RecordBatch;
 import org.apache.kafka.common.utils.Time;
-import org.apache.kafka.server.log.storage.FetchPartitionData;
 import org.apache.kafka.server.share.GroupTopicPartitionData;
 import org.apache.kafka.server.share.PartitionAllData;
 import org.apache.kafka.server.share.PartitionErrorData;
@@ -45,6 +44,7 @@ import org.apache.kafka.server.share.ReadShareGroupStateParameters;
 import org.apache.kafka.server.share.TopicData;
 import org.apache.kafka.server.share.WriteShareGroupStateParameters;
 import org.apache.kafka.server.share.acknowledge.ShareAcknowledgementBatch;
+import org.apache.kafka.server.storage.log.FetchPartitionData;
 import org.apache.kafka.server.util.timer.Timer;
 import org.apache.kafka.server.util.timer.TimerTask;
