+++ core/src/test/java/kafka/server/share/SharePartitionTest.java
@@ -37,7 +37,6 @@ import org.apache.kafka.common.record.MemoryRecordsBuilder;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Time;
-import org.apache.kafka.server.log.storage.FetchPartitionData;
 import org.apache.kafka.server.share.NoOpShareStatePersister;
 import org.apache.kafka.server.share.PartitionFactory;
 import org.apache.kafka.server.share.Persister;
@@ -46,6 +45,7 @@ import org.apache.kafka.server.share.ReadShareGroupStateResult;
 import org.apache.kafka.server.share.TopicData;
 import org.apache.kafka.server.share.WriteShareGroupStateResult;
 import org.apache.kafka.server.share.acknowledge.ShareAcknowledgementBatch;
+import org.apache.kafka.server.storage.log.FetchPartitionData;
 import org.apache.kafka.server.util.FutureUtils;
 import org.apache.kafka.server.util.timer.SystemTimer;
 import org.apache.kafka.server.util.timer.SystemTimerReaper;
