+++ server-common/src/main/java/org/apache/kafka/server/storage/log/FetchPartitionData.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.server.log.storage;
+package org.apache.kafka.server.storage.log;
 
 import org.apache.kafka.common.message.FetchResponseData;
 import org.apache.kafka.common.protocol.Errors;
